docs: document `path` module, less nesting
