+++ packages/vexide-core/src/fs/fs_str.rs
@@ -1,4 +1,3 @@
-use alloc::{borrow::ToOwned, boxed::Box};
 use core::ptr;
 
 #[repr(transparent)]
