+++ packages/vexide-core/src/lib.rs
@@ -23,3 +23,4 @@ pub mod io;
 pub mod program;
 pub mod sync;
 pub mod time;
+pub mod path;
